working on populate, still not working :(
I've tried to update populate using asyc notation, but it looks like the version of node they are using doesnt support that. I tried updating, but that broke a bunch of dependencies (of course). I'll have to look into what one does when stuck in node.js version 6 instead of version 10